+++ source/module_hamilt_lcao/module_gint/kernels/cuda/gint_vl.cuh
@@ -17,7 +17,7 @@ namespace GintKernel{
 
 __global__ void get_psi_and_vldr3(double *ylmcoef,
                                   double delta_r_g,
-                                  double bxyz_g,
+                                  int bxyz_g,
                                   double nwmax_g,
                                   double *input_double,
                                   int *input_int,
